Update LogView.java
There are a lot of Android dependencies here. 

http://developer.android.com/reference/org/apache/http/client/HttpClient.html

Also, do you think we should document where the hardcoded config for this app is somewhere? Took me a minute or two to find it here in public void setConfig, and it may take longer for someone with no context, especially given the filename. 

We're eventually going to want tot change the "send_limit" to "100" by default, too.